feat: abstracted a ExerciseForm that can both handle creating and updating exercises and is very nicely typed :)
